Issue #18: Move .vue files to components dir
Rendering .vue under sub-directories of components folder doesn't
work on MacOS. Move .vue files to components dir and adjust router.

it fixes #18.
